requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
pygments==2.19.1
    # via rich
pyyaml==6.0.1
    # via pre-commit-starter (pyproject.toml)
rich==13.7.0
    # via pre-commit-starter (pyproject.toml)
